LEARNER-6445, fixed comments, code improvement
