+++ Source/CourseDashboardViewController.swift
@@ -195,26 +195,7 @@ class CourseDashboardViewController: UITabBarController, UITabBarControllerDeleg
             present(controller, animated: true, completion: nil)
         }
     }
-    
-    func tabBarViewControllerIndex(with controller: AnyClass, courseOutlineMode: CourseOutlineMode? = .full) -> Int {
-        guard let viewControllers = viewControllers else {
-            return 0
-        }
-    
-        for i in 0..<viewControllers.count {
-            if viewControllers[i].isKind(of: controller) {
-                if  viewControllers[i].isKind(of: CourseOutlineViewController.self)  {
-                    if let viewController = viewControllers[i] as? CourseOutlineViewController, viewController.courseOutlineMode == courseOutlineMode {
-                        return i
-                    }
-                } else {
-                    return i
-                }
-            }
-        }
-        return 0
-    }
-    
+        
 // MARK: Deep Linking
     func switchTab(with type: DeepLinkType) {
         switch type {
